trying out omitting delay
