+++ project1/source/Motor_Control_Component_Test.c
@@ -7,27 +7,17 @@
 
 #include "Motor_Control_Component_Test.h"
 
-SpeedMode_t modes[] = {MODE0, MODE1, MODE2};
-SpeedDirection_t dirs[] = {FORWARD, BACKWARD};
-
-int modes_i, dirs_i;
+int speed;
 void test_MotorTask(){
 
 	TimerHandle_t timer_handle = xTimerCreate("Periodic timer",
-		100 / portTICK_PERIOD_MS,
+		200 / portTICK_PERIOD_MS,
 		pdTRUE, // Indicates periodic
 		NULL,
 		testTimer_MotorTask
 	);
 
-	MotorQueueMessage_t *message = malloc(sizeof(MotorQueueMessage_t));
-	message->type = SPEED_MODE;
-
-	SpeedModePayload_t *payload = malloc(sizeof(SpeedModePayload_t));
-	message->payload = (void*) payload;
-
-	modes_i = 0; dirs_i = 0;
-
+	speed = -90;
 	BaseType_t status = xTimerStart(timer_handle, 0);
 	if(status != pdPASS)
 	{
@@ -38,15 +28,43 @@ void test_MotorTask(){
 
 
 void testTimer_MotorTask(TimerHandle_t handle){
-	SpeedModePayload_t payload = {dirs[dirs_i], modes[modes_i]};
-	MotorQueueMessage_t msg = {SPEED_MODE, (void *) &payload};
+	int * new_speed_p = malloc(sizeof(int));
+	*new_speed_p = speed;
+	PRINTF("[TEST - Motor Task] speed: %d\r\n", speed);
+	xQueueSendToBack(motor_queue, (void*) new_speed_p, portMAX_DELAY);
 
-	PRINTF("[TEST - Motor Task] Mode Code: %d, Direction Code: %d\n", modes[modes_i], dirs[dirs_i]);
-	payload.mode = modes[modes_i];
-	payload.direction = dirs[dirs_i];
+	speed += 10;
+	if(speed == 100)
+		speed = -90;
 
-	xQueueSendToBack(motor_queue, (void*) &msg, portMAX_DELAY);
+}
 
-	modes_i = (modes_i+1)%3;
-	dirs_i = (dirs_i+1)%2;
+int angle;
+void test_PositionTask(){
+
+	TimerHandle_t timer_handle = xTimerCreate(
+		"Position Testing Periodic Timer",
+		200 / portTICK_PERIOD_MS,
+		pdTRUE,
+		NULL,
+		testTimer_PositionTask
+	);
+
+	angle = -45;
+
+	if((xTimerStart(timer_handle, 0)) != pdPASS){
+		PRINTF("Couldn't start the timer\r\n");
+		while(1);
+	}
+}
+
+void testTimer_PositionTask(TimerHandle_t handle){
+	int *new_angle_ptr = malloc(sizeof(int));
+	*new_angle_ptr = angle;
+	PRINTF("[TEST - Position Task] angle: %d\r\n", angle);
+	xQueueSendToBack(angle_queue, (void*) new_angle_ptr, portMAX_DELAY);
+
+	angle += 10;
+	if(angle > 45)
+		angle = -45;
 }
